+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <flutter_nsd/flutter_nsd_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  FlutterNsdPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterNsdPlugin"));
+}